+++ src/javaScript/Auth.js
@@ -1,13 +1,17 @@
-import {useEffect} from "react";
+import {useEffect, useLayoutEffect} from "react";
 import axios from "axios";
 import jwt_decode from "jwt-decode";
 import Cookies from 'js-cookie';
+import {useNavigate} from "react-router-dom";
 
-export default function(SpecificComponent, option){
+export default function(SpecificComponent, option, authorization = null){
+    const navigate = useNavigate();
+    const goToMain = () => {
+        navigate("/");
+    }
     const jwtCheck = async () => {
-        let response = null;
         try {
-            response = await axios.get("http://localhost:8585/jwt/user", {
+            const response = await axios.get("http://localhost:8585/jwt/user", {
                 headers: {
                     Authorization: Cookies.get("jwt")
                 }
@@ -19,12 +23,33 @@ export default function(SpecificComponent, option){
     }
     function AuthenticationCheck(props) {
         useEffect(() => {
+            console.log('Component has rendered');
+            // const checkJwt = async () => {
+            //     const isJwtValid = await jwtCheck() === 200 ? true : false;
+            //     console.log(isJwtValid);
+            //     if (!isJwtValid) {
+            //         if (option) {
+            //             //alert('로그인 안했는데 인증페이지?');
+            //             navigate("/login");
+            //         }
+            //     }else{
+            //
+            //     }
+            //     if (authorization !== null){
+            //
+            //     }
+            // };
+            // checkJwt();
+        }, []);
+        useLayoutEffect(() => {
             const checkJwt = async () => {
+                console.log('Component will render soon');
                 const isJwtValid = await jwtCheck() === 200 ? true : false;
                 console.log(isJwtValid);
                 if (!isJwtValid) {
                     if (option) {
-                        alert('로그인 안했는데 인증페이지?');
+                        //alert('로그인 안했는데 인증페이지?');
+                        navigate("/login");
                     }
                 }
             };
