+++ src/javaScript/page/MyPage.js
@@ -1,10 +1,10 @@
 import {useNavigate} from "react-router-dom";
 
 export default function MyPage(){
-    const navigate = useNavigate();
-    const goToMain = () => {
-        navigate("/");
-    }
+    // const navigate = useNavigate();
+    // const goToMain = () => {
+    //     navigate("/");
+    // }
     return (
         <div>
             <p>마이페이지 입니다.</p>
